Improvement for [HER-1] Support 'Crawl-Delay' and 'Allow' robots.txt directives * AbstractFrontier.java     change boolean 'respect-crawl-delay' to integer 'respect-crawl-delay-up-to-secs' Fix for [HER-1516] in datamodel.Robotstxt, Integer.parseInt generates NumberFormatException on floating point Crawl-delay strings * RobotsDirectives.java, RobotsTxt.java, RobotsExclusionPolicy.java     parse, store, return crawlDelay as float * RobotsTxtTest.java     compare value as float

+++ modules/src/main/java/org/archive/modules/net/RobotsDirectives.java
@@ -37,7 +37,7 @@ public class RobotsDirectives implements Serializable {
     
     PrefixSet disallows = new PrefixSet();
     PrefixSet allows = new PrefixSet();
-    int crawlDelay = -1; 
+    float crawlDelay = -1; 
 
     public boolean allows(String path) {
         if(disallows.containsPrefixOf(path)) {
@@ -59,11 +59,11 @@ public class RobotsDirectives implements Serializable {
         allows.add(path);
     }
 
-    public void setCrawlDelay(int i) {
+    public void setCrawlDelay(float i) {
         crawlDelay=i;
     }
 
-    public int getCrawlDelay() {
+    public float getCrawlDelay() {
         return crawlDelay;
     }
 }
